update base image repository - split; add remote image repository to figma-remote
