+++ figma-remote/lib/asset-repository/index.ts
@@ -1,0 +1,1 @@
+export * from "./image-repository";
