+++ figma-remote/lib/index.ts
@@ -1,3 +1,4 @@
 export * as types from "@design-sdk/figma-remote-types";
 export * as mapper from "./mapper";
 export * as fetch from "./fetch";
+export { configure_auth_credentials } from "./configure-auth-credentials";
